Phase clock for Puredata
This patch is conceived as a clock, using phasing information to tell the time. 12 harmonics of a base note are set off at slightly different tempos such that they realign after one hour. The effect is that consonant intervals of the hour align the notes (eg. at quarter past, you will hear four distinct beeps per second) while in between there is somewhat chaotic rhythms. 

The pitch slowly rises throughout the hour, moving up one semitone per hour, and jumping back down an octave at 12:00. My theory in creating this was that an individual who grew up with this kind of clock would develop a very unique sensibility of tempo/rhythm and pitch (imagine drawing a strong correlation between the note F# and bedtime, for instance, or being able to accurately divide a second into 7 notes as a tempo reference.)

I would like to acknowledge Steve Reich, who was an obvious inspiration for this project.